+++ src/main/java/com/evstore/ecommerce/repository/CartRepository.java
@@ -8,11 +8,10 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import com.evstore.ecommerce.model.Cart;
-import com.evstore.ecommerce.user.User;
-
+import com.evstore.ecommerce.model.User;
 
 @Repository
-public interface CartRepository extends JpaRepository<Cart,Long>{
+public interface CartRepository extends JpaRepository<Cart, Integer> {
   
     Optional<Cart> findByUser(User user);
 }
